requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-dotenv>=1.0.0
cohere>=4.1.0
